app/requirements.txt
psycopg2~=2.9.9
flask~=3.0.2
bcrypt~=4.1.2
python-dotenv~=1.0.1
pyjwt~=2.8.0
pyOpenSSL==24.1.0
Flask-JWT-Extended==4.6.0
flask-cors==4.0.1
Flask-WTF==1.2.1